Autenticación de usuarios, acceso de rutas, manejo de scope

+++ Backend/src/routes/user.js
@@ -2,12 +2,14 @@ const express = require("express");
 const userSchema = require("../models/user");
 const bcrypt = require("bcryptjs");
 const createAccessToken = require("../libs/jwt");
+const authRequerid = require("../middlewares/validateTokenAdmin");
 
 const router = express.Router();
 
 //Create users
 router.post("/register", async (req, res) => {
   const { name, email, password } = req.body;
+
   try {
     const userExist = await userSchema.findOne({ email });
     if (userExist)
@@ -22,8 +24,8 @@ router.post("/register", async (req, res) => {
     user.save();
 
     const token = await createAccessToken({ id: user._id });
-    res.cookie("token", token),
-      res.json({ message: `User created succesfully` });
+    res.cookie("token", token);
+    res.json({ message: `User created succesfully` });
   } catch (err) {
     res.send(`Error: ${err}`);
   }
@@ -39,13 +41,30 @@ router.post("/login", async (req, res) => {
 
     if (!match) return res.status(400).json({ message: "Wrong password" });
 
-    const token = await createAccessToken({ id: userFound._id });
-    res.cookie("token", token),
-      res.json({
+    if (userFound.admin) {
+      const token = await createAccessToken({
         id: userFound._id,
-        name: userFound.name,
-        email: userFound.email,
+        scope: "admin",
       });
+      res.cookie("token", token),
+        res.json({
+          id: userFound._id,
+          name: userFound.name,
+          email: userFound.email,
+          admin: userFound.admin,
+        });
+    } else {
+      const token = await createAccessToken({
+        id: userFound._id,
+        scope: "user",
+      });
+      res.cookie("token", token),
+        res.json({
+          id: userFound._id,
+          name: userFound.name,
+          email: userFound.email,
+        });
+    }
   } catch (err) {
     res.send(err);
   }
@@ -54,6 +73,19 @@ router.post("/login", async (req, res) => {
 router.post("/logout", (req, res) => {
   res.cookie("token", "", { expires: new Date(0) });
   return res.sendStatus(200);
+});
+
+router.get("/profile", authRequerid, async (req, res) => {
+  const userFound = await userSchema.findById(req.user.id);
+
+  if (!userFound) return res.status(401).json({ message: "User not found" });
+
+  res.json({
+    id: userFound._id,
+    name: userFound.name,
+    email: userFound.email,
+    admin: userFound.admin,
+  });
 });
 
 // Get all users
@@ -84,12 +116,16 @@ router.put("/users/:id", (req, res) => {
 });
 
 // Delete a User
-router.delete("/users/:id", (req, res) => {
-  const { id } = req.params;
-  userSchema
-    .findByIdAndRemove(id)
-    .then((data) => res.json(data))
-    .catch((error) => res.json({ message: error }));
+router.delete("/users/:id", authRequerid, (req, res) => {
+  if (req.admin) {
+    const { id } = req.params;
+    userSchema
+      .findByIdAndRemove(id)
+      .then((data) => res.json(data))
+      .catch((error) => res.json({ message: error }));
+  } else {
+    res.status(401).json({ message: "User not authorizated" });
+  }
 });
 
 module.exports = router;
